fix:🐛routeのクロージャーのコメントアウト

+++ routes/web.php
@@ -47,7 +47,7 @@ Route::get('/profile/index','ProfileController@index')->name('profile.index');
 //プロフィール編集
 Route::get('/profile/{user}','ProfileController@edit')->name('profile.edit');
 Route::put('/profile/{user}','ProfileController@update')->name('profile.update');
-//管理者
+//管理者 ->herokuエラー。クロージャーは使えないため、上記記述を変更
 // Route::middleware(['can:admin'])->group(function(){
 //     Route::get('/profile/index', 'ProfileController@index')->name('profile.index');
 //     Route::delete('/profile/delete/{user}', 'ProfileController@delete')->name('profile.delete');
@@ -55,9 +55,9 @@ Route::put('/profile/{user}','ProfileController@update')->name('profile.update')
 //     Route::put('/roles/{user}/detach', 'RoleController@detach')->name('role.detach');
 // });
 Route::get('/profile/index', 'ProfileController@index')->name('profile.index')->middleware(['can:admin']);
-Route::delete('/profile/delete/{user}', 'ProfileController@delete')->name('profile.delete');
-Route::put('/roles/{user}/attach', 'RoleController@attach')->name('role.attach');
-Route::put('/roles/{user}/detach', 'RoleController@detach')->name('role.detach');
+Route::delete('/profile/delete/{user}', 'ProfileController@delete')->name('profile.delete')->middleware(['can:admin']);
+Route::put('/roles/{user}/attach', 'RoleController@attach')->name('role.attach')->middleware(['can:admin']);
+Route::put('/roles/{user}/detach', 'RoleController@detach')->name('role.detach')->middleware(['can:admin']);
 //お気に入り機能
 Route::get('/reply/favorite/{incidentPost}', 'FavoriteController@favorite')->name('favorite');
 Route::get('/reply/unfavorite/{incidentPost}', 'FavoriteController@unfavorite')->name('unfavorite');
